Finalizadas implementaciones de las consultas

+++ src/main/java/edu/eci/cvds/samples/services/ServiciosAlquiler.java
@@ -9,7 +9,7 @@ import edu.eci.cvds.samples.entities.ItemRentado;
 import edu.eci.cvds.samples.entities.TipoItem;
 
 public interface ServiciosAlquiler {
-    public abstract int valorMultaRetrasoxDia(int itemId);
+    public abstract long valorMultaRetrasoxDia(int itemId);
 
     public abstract Cliente consultarCliente(long docu) throws ExcepcionServiciosAlquiler;
