+++ src/main/java/edu/eci/cvds/samples/services/impl/ServiciosAlquilerStub.java
@@ -38,7 +38,7 @@ public class ServiciosAlquilerStub implements ServiciosAlquiler {
     }
 
     @Override
-    public int valorMultaRetrasoxDia(int itemId) {
+    public long valorMultaRetrasoxDia(int itemId) {
         return MULTA_DIARIA;
     }
